Implemented myc as in Gerris. Reorganiseed the Test directory. Fixed minor changes.
Added some README files in the documentation.

+++ Tests/clean.sh
@@ -1,11 +1,14 @@
 #!/bin/bash
 #set -x
 
+rm -fR *~ *.tmp tmp*
 
 for dir in `ls`; do 
-    cd $dir
-    echo "in dir " $dir
-    echo "rm -fR *.o *.mod paris stats *~ track out* errftc tmp* *.tmp fort.* *.visit core.*"
-    rm -fR *.o *.mod paris stats *~ track out* errftc tmp* *.tmp fort.* *.visit core.*
-    cd ..
+    if [ -d $dir ]; then
+	cd $dir
+#	echo "in dir " `pwd`
+#	echo "rm -fR *.o *.mod paris stats *~ track out* errftc tmp* *.tmp fort.* *.visit core.*"
+	rm -fR *.o *.mod paris stats *~ track out* errftc tmp* *.tmp fort.* *.visit core.*
+	cd ..
+    fi
 done
